[analyzer] Find comment references in DocImportScopes.
This CL adds the `DocImportScope` which allows the scope visitor to find elements in libraries imported by a `@docImport`. The resolved elements will eventually be used in Dartdoc to make links for the comment references.

`Scope.lookup` can choose toggle whether or not it needs to look in doc imports for a certain id.

Change-Id: I329152651ea3cb4e3a5e2787a5d91ba1cd397d29
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/353232
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Bob Nystrom <rnystrom@google.com>