[analysis_server] Change file URI labels in completions that add imports to relative paths in new details
This fixes part of https://github.com/dart-lang/sdk/issues/55013 (which is that we'd display full file URIs as the imported library on completions when they're not a package URI).

There's still some other parts to this issue (such as picking this completion up as an importing completion when it's in the same file) but I'll do that in another change.

Change-Id: Id14f7797bfd98621ac6bbb71addd52f928c6c0ff
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/355141
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Keerti Parthasarathy <keertip@google.com>
Commit-Queue: Keerti Parthasarathy <keertip@google.com>