[analysis_server] Remove forced delays in sending requests to LSP server in tests
I don't recall the reason for these Futures, because the streams are already async anyway (we don't pass `sync: true` to the constructors). These were causing issues in a new test I'm trying to write (which needs to send two requests together and ensure they're handled correctly, but these futures were adding a delay between them that I didn't want).

Removing these broke one test which I suspect was flaky because it was relying on some requests to overlap just based on this timing. I have changed that test to use a Future (that we added for some similar tests) to verify the same thing reliably.

Change-Id: I031c785dff5d76ed61109ce6beb26b66f7276c20
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/356721
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>