[analysis_server] Handle requests for resolved units in handlers between initialization and analysis roots being set
If you open VS Code and have some Dart files open, it may send requests to the analysis server before we have (asynchronously) set up the analysis roots. This often results in "File not analyzed" errors in the console, but until now I had believed these errors to not cause any visible issues.

However, https://github.com/Dart-Code/Dart-Code/issues/5030 had a good repro for where this is visible - the Outline view would never populate for these open files.

We had some existing code to wait for in-progress context rebuilds to complete when fetching a resolved unit in a request handler, but requests could still get in before that started.

This change also waits for the client to send Initialized, which is when we start the analysis root build, so this bridges that gap and ensures the initial analysis roots are set before any handlers get to run.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5030

Change-Id: I4b2d936f533f0696e3e363972faa82a2304276c0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/356722
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>