[analysis_server] Add new LSP commands for (Preview/) Fix All in Workspace
This adds two new commands:

- Preview All Fixes in Workspace
- Apply All Fixes in Workspace

These use the BulkFixProcessor (like "dart fix" would) to produce edits to be run in the IDE. The implementation of both commands is the same, but one sets the "needs confirmation" flag on the annotated edits so that the user will see them in the editor rather than them be applied immediately.

These won't show up automatically, the client (Dart-Code) will need to register them with VS Code so they show up where the user can execute them (which will be done only if the server advertises these commands, and only for an SDK version when we determine this is ready).

This does not yet update the prompt that users see if they open a workspace that's "dart fix"able (but I'd like to also do that when we're happy this is good).

Change-Id: Ida15cc2209b21493a7606a70c754cf25aee29ed9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/356161
Commit-Queue: Keerti Parthasarathy <keertip@google.com>
Reviewed-by: Keerti Parthasarathy <keertip@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>