[analysis_server] Add the ability for the "dart fix" prompt to use in-editor application of all fixes
This is currently behind an `initializationOption` supplied by the client (so that in the short-term we can control this behaviour without SDK changes, for example deciding to enable this in the next stable without having to decide before the branch).

When conditions are met (the client flag, and applyEdit/changeAnnotations) instead of showing a prompt pointing the user to the website to learn about "dart fix" give buttons for running them in-editor. An example of how the prompt looks is here:

https://github.com/Dart-Code/Dart-Code/issues/5022#issuecomment-1991562705

Initially, the new commands and this client flag will be gated on an experimental flag to allow wider testing before it starts showing up.

Change-Id: I64254162e8d146ebf4075bade37b2efef76e11c4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/357120
Reviewed-by: Keerti Parthasarathy <keertip@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Keerti Parthasarathy <keertip@google.com>