[analysis_server] Fix handling of LSP completion display text with spaces
For LSP we sometimes split the display text to populate a number of different fields (for example when the display text is `foo => bar` we only want `foo` in the `filterText`).

We previously split on spaces but that fails for the new `override foo()` display texts, so this was changed to split on equals instead (and trim to remove any trailing spaces).

Fixes https://github.com/dart-lang/sdk/issues/32677

Change-Id: I7494426cfcc95388f0f1f91f7f9ff7bcac0f23ca
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/357320
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>